Signals and alerts now working
